poker-backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for poker simulation app"
authors = ["biftu <biftu@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.13"
fastapi = ">=0.115.2,<0.116.0"
uvicorn = { extras = ["standard"], version = ">=0.34.2,<0.35.0" }
psycopg2-binary = ">=2.9.10,<3.0.0"
pydantic = ">=2.11.0,<3.0.0"
pokerkit = ">=0.6.3,<0.7.0"
python-dotenv = ">=1.1.0,<2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"
ruff = "^0.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"